add txn_name to promise_type via handler_t::awaitable
a little hacky perhaps. but it improves logging (we always know what
txn we're in, within promise code), and, more importantly, it simplifies
calling the dp::txn::complete() overloads, which are now shared (in dp::).

put the starting touches on making sellitem::txn:handler a proper loop,
but #ifdef'd it out for the purpose of this checkpoint.